+++ lib/rubocop/cop/style/block_end_newline.rb
@@ -4,7 +4,7 @@
 module RuboCop
   module Cop
     module Style
-      # This cop checks whether the end statement of a do end blocks
+      # This cop checks whether the end statement of a do..end block
       # is on its own line.
       #
       # @example